9-yard-greening

+++ 1-cpp-basics/1-first-steps-in-coding/lab/9-yard-greening.cpp
@@ -6,7 +6,7 @@ int main() {
 	const double DISCOUNT_PERCENT = 0.18;
 	double totalPrice, discount;
 	
-	int yardSize;
+	double yardSize;
 	cin >> yardSize;
 	
 	totalPrice = yardSize * YARD_GREENING_PRICE;
